refactor: move some type-related utils into module
