+++ src/student_success_tool/utils.py
@@ -1,27 +1,12 @@
 import logging
 import re
 import typing as t
-from collections.abc import Collection, Iterable
+from collections.abc import Iterable
 
 LOGGER = logging.getLogger(__name__)
 
 RE_VARIOUS_PUNCTS = re.compile(r"[!()*+\,\-./:;<=>?[\]^_{|}~]")
 RE_QUOTATION_MARKS = re.compile(r"[\'\"\`]")
-
-
-def to_list(value: t.Any) -> list:
-    """Cast ``value`` into a list, regardless of its type."""
-    if isinstance(value, list):
-        return value
-    elif isinstance(value, Iterable) and not isinstance(value, (str, bytes)):
-        return list(value)
-    else:
-        return [value]
-
-
-def is_collection_but_not_string(value: t.Any) -> bool:
-    """Check if ``value`` is a collection, excluding strings (and bytes)."""
-    return isinstance(value, Collection) and not isinstance(value, (str, bytes))
 
 
 def unique_elements_in_order(eles: Iterable) -> Iterable:
